v201105: updating extract_features and extract_embeddings
