+++ scripts/extract_embeddings.py
